+++ viennacl/tools/sha1.hpp
@@ -117,30 +117,30 @@ namespace detail
       digest32_t d32;
       getDigest(d32);
       size_t di = 0;
-      digest[di++] = ((d32[0] >> 24) & 0xFF);
-      digest[di++] = ((d32[0] >> 16) & 0xFF);
-      digest[di++] = ((d32[0] >> 8) & 0xFF);
-      digest[di++] = ((d32[0]) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[0] >> 24) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[0] >> 16) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[0] >> 8) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[0]) & 0xFF);
 
-      digest[di++] = ((d32[1] >> 24) & 0xFF);
-      digest[di++] = ((d32[1] >> 16) & 0xFF);
-      digest[di++] = ((d32[1] >> 8) & 0xFF);
-      digest[di++] = ((d32[1]) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[1] >> 24) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[1] >> 16) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[1] >> 8) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[1]) & 0xFF);
 
-      digest[di++] = ((d32[2] >> 24) & 0xFF);
-      digest[di++] = ((d32[2] >> 16) & 0xFF);
-      digest[di++] = ((d32[2] >> 8) & 0xFF);
-      digest[di++] = ((d32[2]) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[2] >> 24) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[2] >> 16) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[2] >> 8) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[2]) & 0xFF);
 
-      digest[di++] = ((d32[3] >> 24) & 0xFF);
-      digest[di++] = ((d32[3] >> 16) & 0xFF);
-      digest[di++] = ((d32[3] >> 8) & 0xFF);
-      digest[di++] = ((d32[3]) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[3] >> 24) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[3] >> 16) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[3] >> 8) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[3]) & 0xFF);
 
-      digest[di++] = ((d32[4] >> 24) & 0xFF);
-      digest[di++] = ((d32[4] >> 16) & 0xFF);
-      digest[di++] = ((d32[4] >> 8) & 0xFF);
-      digest[di++] = ((d32[4]) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[4] >> 24) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[4] >> 16) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[4] >> 8) & 0xFF);
+      digest[di++] = static_cast<uint8_t>((d32[4]) & 0xFF);
       return digest;
     }
